fix(appbar): apply material custom appbar to each screen
